requirements.txt
telethon
feedparser
pyyaml
beautifulsoup4
aiohttp
tortoise-orm
durations_nlp

# db drivers
aiosqlite
asyncpg
aiomysql

# for importing data from previous versions
msgpack